feat: Move captureLastError to SDK
